Improvements to signIn - capital letters in username is turn to lowercase; getElucidation seems to work

+++ src/main/java/ACQ/IDialog.java
@@ -11,7 +11,7 @@ public interface IDialog {
      * Create a new meeting and add it to the collection of meetings
      * @return new IMeeting
      */
-    IMeeting createMeeting(IUser currentUser, long meetingNumber);
+    IMeeting createMeeting(IUser currentUser, int meetingNumber);
 
     /**
      * Cancel a meeting, and remove it from the collection of meetings
